+++ egui/src/context.rs
@@ -639,7 +639,7 @@ impl Context {
     /// Will become active at the start of the next frame.
     ///
     /// Note that this may be overwritten by input from the integration via [`RawInput::pixels_per_point`].
-    /// For instance, when using `egui_web` the browsers native zoom level will always be used.
+    /// For instance, when using `eframe` on web, the browsers native zoom level will always be used.
     pub fn set_pixels_per_point(&self, pixels_per_point: f32) {
         if pixels_per_point != self.pixels_per_point() {
             self.request_repaint();