+++ egui/src/ui.rs
@@ -1513,8 +1513,8 @@ impl Ui {
 
     /// Show an image here with the given size.
     ///
-    /// In order to display an image you must first acquire a [`TextureHandle`]
-    /// using [`Context::load_texture`].
+    /// In order to display an image you must first acquire a [`TextureHandle`].
+    /// This is best done with [`egui_extras::RetainedImage`](https://docs.rs/egui_extras/latest/egui_extras/image/struct.RetainedImage.html) or [`Context::load_texture`].
     ///
     /// ```
     /// struct MyImage {