+++ egui_demo_app/src/wrap_app.rs
@@ -1,3 +1,6 @@
+use egui_demo_lib::is_mobile;
+use egui_glow::glow;
+
 #[derive(Default)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 struct EasyMarkApp {
@@ -41,6 +44,7 @@ impl eframe::App for FractalClockApp {
             });
     }
 }
+
 // ----------------------------------------------------------------------------
 
 #[derive(Default)]
@@ -67,174 +71,234 @@ impl eframe::App for ColorTestApp {
 
 // ----------------------------------------------------------------------------
 
-/// All the different demo apps.
+/// The state that we persist (serialize).
 #[derive(Default)]
 #[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
 #[cfg_attr(feature = "serde", serde(default))]
-pub struct Apps {
+pub struct State {
     demo: DemoApp,
     easy_mark_editor: EasyMarkApp,
     #[cfg(feature = "http")]
     http: crate::apps::HttpApp,
     clock: FractalClockApp,
     color_test: ColorTestApp,
+
+    selected_anchor: String,
+    backend_panel: super::backend_panel::BackendPanel,
 }
 
-impl Apps {
-    fn iter_mut(&mut self) -> impl Iterator<Item = (&str, &str, &mut dyn eframe::App)> {
+/// Wraps many demo/test apps into one.
+pub struct WrapApp {
+    state: State,
+    // not serialized (because it contains OpenGL buffers etc)
+    custom3d: crate::apps::Custom3d,
+    dropped_files: Vec<egui::DroppedFile>,
+}
+
+impl WrapApp {
+    pub fn new(cc: &eframe::CreationContext<'_>) -> Self {
+        #[allow(unused_mut)]
+        let mut slf = Self {
+            state: State::default(),
+            custom3d: crate::apps::Custom3d::new(&cc.gl),
+            dropped_files: Default::default(),
+        };
+
+        #[cfg(feature = "persistence")]
+        if let Some(storage) = cc.storage {
+            if let Some(state) = eframe::get_value(storage, eframe::APP_KEY) {
+                slf.state = state;
+            }
+        }
+
+        if cc.integration_info.prefer_dark_mode == Some(false) {
+            cc.egui_ctx.set_visuals(egui::Visuals::light()); // use light mode if explicitly asked for
+        } else {
+            cc.egui_ctx.set_visuals(egui::Visuals::dark()); // use dark mode if there is no preference, or the preference is dark mode
+        }
+
+        slf
+    }
+
+    fn apps_iter_mut(&mut self) -> impl Iterator<Item = (&str, &str, &mut dyn eframe::App)> {
         vec![
-            ("✨ Demos", "demo", &mut self.demo as &mut dyn eframe::App),
+            (
+                "✨ Demos",
+                "demo",
+                &mut self.state.demo as &mut dyn eframe::App,
+            ),
             (
                 "🖹 EasyMark editor",
                 "easymark",
-                &mut self.easy_mark_editor as &mut dyn eframe::App,
+                &mut self.state.easy_mark_editor as &mut dyn eframe::App,
             ),
             #[cfg(feature = "http")]
-            ("⬇ HTTP", "http", &mut self.http as &mut dyn eframe::App),
+            (
+                "⬇ HTTP",
+                "http",
+                &mut self.state.http as &mut dyn eframe::App,
+            ),
             (
                 "🕑 Fractal Clock",
                 "clock",
-                &mut self.clock as &mut dyn eframe::App,
+                &mut self.state.clock as &mut dyn eframe::App,
+            ),
+            (
+                "🔺 3D painting",
+                "custom3e",
+                &mut self.custom3d as &mut dyn eframe::App,
             ),
             (
                 "🎨 Color test",
                 "colors",
-                &mut self.color_test as &mut dyn eframe::App,
+                &mut self.state.color_test as &mut dyn eframe::App,
             ),
         ]
         .into_iter()
     }
 }
 
-/// Wraps many demo/test apps into one.
-#[derive(Default)]
-#[cfg_attr(feature = "serde", derive(serde::Deserialize, serde::Serialize))]
-#[cfg_attr(feature = "serde", serde(default))]
-pub struct WrapApp {
-    selected_anchor: String,
-    apps: Apps,
-    backend_panel: super::backend_panel::BackendPanel,
-    #[cfg_attr(feature = "serde", serde(skip))]
-    dropped_files: Vec<egui::DroppedFile>,
-}
-
-impl WrapApp {
-    pub fn new(_cc: &eframe::CreationContext<'_>) -> Self {
-        #[cfg(feature = "persistence")]
-        if let Some(storage) = _cc.storage {
-            return eframe::get_value(storage, eframe::APP_KEY).unwrap_or_default();
-        }
-        Self::default()
-    }
-}
-
 impl eframe::App for WrapApp {
     #[cfg(feature = "persistence")]
     fn save(&mut self, storage: &mut dyn eframe::Storage) {
-        eframe::set_value(storage, eframe::APP_KEY, self);
+        eframe::set_value(storage, eframe::APP_KEY, &self.state);
     }
 
-    fn clear_color(&self) -> egui::Rgba {
-        egui::Rgba::TRANSPARENT // we set a [`CentralPanel`] fill color in `demo_windows.rs`
+    fn clear_color(&self, visuals: &egui::Visuals) -> egui::Rgba {
+        visuals.window_fill().into()
     }
 
     fn update(&mut self, ctx: &egui::Context, frame: &mut eframe::Frame) {
         if let Some(web_info) = frame.info().web_info.as_ref() {
             if let Some(anchor) = web_info.location.hash.strip_prefix('#') {
-                self.selected_anchor = anchor.to_owned();
+                self.state.selected_anchor = anchor.to_owned();
             }
         }
 
-        if self.selected_anchor.is_empty() {
-            self.selected_anchor = self.apps.iter_mut().next().unwrap().0.to_owned();
+        if self.state.selected_anchor.is_empty() {
+            let selected_anchor = self.apps_iter_mut().next().unwrap().0.to_owned();
+            self.state.selected_anchor = selected_anchor;
         }
 
         egui::TopBottomPanel::top("wrap_app_top_bar").show(ctx, |ui| {
             egui::trace!(ui);
-            self.bar_contents(ui, frame);
+            ui.horizontal_wrapped(|ui| {
+                ui.visuals_mut().button_frame = false;
+                self.bar_contents(ui, frame);
+            });
         });
 
-        self.backend_panel.update(ctx, frame);
+        self.state.backend_panel.update(ctx, frame);
 
-        if self.backend_panel.open || ctx.memory().everything_is_visible() {
-            egui::SidePanel::left("backend_panel").show(ctx, |ui| {
-                self.backend_panel.ui(ui, frame);
+        if !is_mobile(ctx)
+            && (self.state.backend_panel.open || ctx.memory().everything_is_visible())
+        {
+            egui::SidePanel::left("backend_panel")
+                .resizable(false)
+                .show(ctx, |ui| {
+                    ui.vertical_centered(|ui| {
+                        ui.heading("💻 Backend");
+                    });
 
-                ui.separator();
-
-                ui.horizontal(|ui| {
-                    if ui
-                        .button("Reset egui")
-                        .on_hover_text("Forget scroll, positions, sizes etc")
-                        .clicked()
-                    {
-                        *ui.ctx().memory() = Default::default();
-                    }
-
-                    if ui.button("Reset everything").clicked() {
-                        *self = Default::default();
-                        *ui.ctx().memory() = Default::default();
-                    }
+                    ui.separator();
+                    self.backend_panel_contents(ui, frame);
                 });
-            });
         }
 
-        let mut found_anchor = false;
+        self.show_selected_app(ctx, frame);
 
-        for (_name, anchor, app) in self.apps.iter_mut() {
-            if anchor == self.selected_anchor || ctx.memory().everything_is_visible() {
-                app.update(ctx, frame);
-                found_anchor = true;
-            }
-        }
-
-        if !found_anchor {
-            self.selected_anchor = "demo".into();
-        }
-
-        self.backend_panel.end_of_frame(ctx);
+        self.state.backend_panel.end_of_frame(ctx);
 
         self.ui_file_drag_and_drop(ctx);
+    }
+
+    fn on_exit(&mut self, gl: &glow::Context) {
+        self.custom3d.on_exit(gl);
     }
 }
 
 impl WrapApp {
+    fn backend_panel_contents(&mut self, ui: &mut egui::Ui, frame: &mut eframe::Frame) {
+        self.state.backend_panel.ui(ui, frame);
+
+        ui.separator();
+
+        ui.horizontal(|ui| {
+            if ui
+                .button("Reset egui")
+                .on_hover_text("Forget scroll, positions, sizes etc")
+                .clicked()
+            {
+                *ui.ctx().memory() = Default::default();
+                ui.close_menu();
+            }
+
+            if ui.button("Reset everything").clicked() {
+                self.state = Default::default();
+                *ui.ctx().memory() = Default::default();
+                ui.close_menu();
+            }
+        });
+    }
+
+    fn show_selected_app(&mut self, ctx: &egui::Context, frame: &mut eframe::Frame) {
+        let mut found_anchor = false;
+        let selected_anchor = self.state.selected_anchor.clone();
+        for (_name, anchor, app) in self.apps_iter_mut() {
+            if anchor == selected_anchor || ctx.memory().everything_is_visible() {
+                app.update(ctx, frame);
+                found_anchor = true;
+            }
+        }
+        if !found_anchor {
+            self.state.selected_anchor = "demo".into();
+        }
+    }
+
     fn bar_contents(&mut self, ui: &mut egui::Ui, frame: &mut eframe::Frame) {
-        // A menu-bar is a horizontal layout with some special styles applied.
-        // egui::menu::bar(ui, |ui| {
-        ui.horizontal_wrapped(|ui| {
-            egui::widgets::global_dark_light_mode_switch(ui);
+        egui::widgets::global_dark_light_mode_switch(ui);
 
-            ui.checkbox(&mut self.backend_panel.open, "💻 Backend");
-            ui.separator();
+        ui.separator();
 
-            for (name, anchor, _app) in self.apps.iter_mut() {
-                if ui
-                    .selectable_label(self.selected_anchor == anchor, name)
-                    .clicked()
-                {
-                    self.selected_anchor = anchor.to_owned();
-                    if frame.is_web() {
-                        ui.output().open_url(format!("#{}", anchor));
+        if is_mobile(ui.ctx()) {
+            ui.menu_button("💻 Backend", |ui| {
+                ui.set_style(ui.ctx().style()); // ignore the "menu" style set by `menu_button`.
+                self.backend_panel_contents(ui, frame);
+            });
+        } else {
+            ui.toggle_value(&mut self.state.backend_panel.open, "💻 Backend");
+        }
+
+        ui.separator();
+
+        let mut selected_anchor = self.state.selected_anchor.clone();
+        for (name, anchor, _app) in self.apps_iter_mut() {
+            if ui
+                .selectable_label(selected_anchor == anchor, name)
+                .clicked()
+            {
+                selected_anchor = anchor.to_owned();
+                if frame.is_web() {
+                    ui.output().open_url(format!("#{}", anchor));
+                }
+            }
+        }
+        self.state.selected_anchor = selected_anchor;
+
+        ui.with_layout(egui::Layout::right_to_left(), |ui| {
+            if false {
+                // TODO: fix the overlap on small screens
+                if let Some(seconds_since_midnight) = crate::seconds_since_midnight() {
+                    if clock_button(ui, seconds_since_midnight).clicked() {
+                        self.state.selected_anchor = "clock".to_owned();
+                        if frame.is_web() {
+                            ui.output().open_url("#clock");
+                        }
                     }
                 }
             }
 
-            ui.with_layout(egui::Layout::right_to_left(), |ui| {
-                if false {
-                    // TODO: fix the overlap on small screens
-                    if let Some(seconds_since_midnight) = crate::seconds_since_midnight() {
-                        if clock_button(ui, seconds_since_midnight).clicked() {
-                            self.selected_anchor = "clock".to_owned();
-                            if frame.is_web() {
-                                ui.output().open_url("#clock");
-                            }
-                        }
-                    }
-                }
-
-                egui::warn_if_debug_build(ui);
-            });
+            egui::warn_if_debug_build(ui);
         });
     }